requirements.txt
dnspython==1.15.0
psutil==5.4.5
python-etcd==0.4.5
tornado==5.0.2
urllib3==1.22
